Started an outlining system
Changing up the operation colors to use outlines instead of tinting the actual graphics.  This should be great because it then it will match the icons we have for each of the matrix operations.  Not finished yet because the outlines will need to be enabled-disabled on operation start and end